Refactor: move util from internal/pkg/ to pkg/
