+++ internal/pkg/plugin/argocd/install.go
@@ -3,11 +3,10 @@ package argocd
 import (
 	"fmt"
 
-	"github.com/merico-dev/stream/internal/pkg/log"
-
 	"github.com/mitchellh/mapstructure"
 
-	"github.com/merico-dev/stream/internal/pkg/util/helm"
+	"github.com/merico-dev/stream/internal/pkg/log"
+	"github.com/merico-dev/stream/pkg/util/helm"
 )
 
 // Install installs ArgoCD with provided options.
