+++ internal/pkg/plugin/argocd/uninstall.go
@@ -3,11 +3,10 @@ package argocd
 import (
 	"fmt"
 
-	"github.com/merico-dev/stream/internal/pkg/log"
-
 	"github.com/mitchellh/mapstructure"
 
-	"github.com/merico-dev/stream/internal/pkg/util/helm"
+	"github.com/merico-dev/stream/internal/pkg/log"
+	"github.com/merico-dev/stream/pkg/util/helm"
 )
 
 func Uninstall(options *map[string]interface{}) (bool, error) {
