+++ internal/pkg/plugin/argocd/validation.go
@@ -1,6 +1,8 @@
 package argocd
 
-import "github.com/merico-dev/stream/internal/pkg/util/helm"
+import (
+	"github.com/merico-dev/stream/pkg/util/helm"
+)
 
 // validate validates the options provided by the core.
 func validate(param *helm.HelmParam) []error {
