+++ internal/pkg/plugin/githubactions/install.go
@@ -4,7 +4,7 @@ import (
 	"github.com/spf13/viper"
 
 	"github.com/merico-dev/stream/internal/pkg/log"
-	"github.com/merico-dev/stream/internal/pkg/util/github"
+	"github.com/merico-dev/stream/pkg/util/github"
 )
 
 // Install sets up GitHub Actions workflows.
