+++ internal/pkg/plugin/kubeprometheus/ishealthy.go
@@ -3,12 +3,11 @@ package kubeprometheus
 import (
 	"fmt"
 
-	"github.com/merico-dev/stream/internal/pkg/log"
-
 	"github.com/mitchellh/mapstructure"
 
-	"github.com/merico-dev/stream/internal/pkg/util/helm"
-	"github.com/merico-dev/stream/internal/pkg/util/k8s"
+	"github.com/merico-dev/stream/internal/pkg/log"
+	"github.com/merico-dev/stream/pkg/util/helm"
+	"github.com/merico-dev/stream/pkg/util/k8s"
 )
 
 const (
