+++ internal/pkg/plugin/reposcaffolding/github/golang/golang.go
@@ -11,7 +11,7 @@ import (
 	"text/template"
 
 	"github.com/merico-dev/stream/internal/pkg/log"
-	"github.com/merico-dev/stream/internal/pkg/util/github"
+	"github.com/merico-dev/stream/pkg/util/github"
 )
 
 const (
