+++ internal/pkg/statemanager/states_map.go
@@ -3,11 +3,10 @@ package statemanager
 import (
 	"bytes"
 
-	"github.com/merico-dev/stream/internal/pkg/log"
-
 	"gopkg.in/yaml.v3"
 
-	"github.com/merico-dev/stream/internal/pkg/container/concurrentmap"
+	"github.com/merico-dev/stream/internal/pkg/log"
+	"github.com/merico-dev/stream/pkg/util/mapz/concurrentmap"
 )
 
 type StatesMap concurrentmap.ConcurrentMap
