+++ pkg/util/github/github_test.go
@@ -6,7 +6,7 @@ import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 
-	"github.com/merico-dev/stream/internal/pkg/util/github"
+	"github.com/merico-dev/stream/pkg/util/github"
 )
 
 var _ = Describe("GitHub", func() {
